Ensure UPDATE_SIGN and TILE_ENTITY_DATA packets are sent correctly.
This avoids the "UNABLE TO LOCATE SIGN" error message and empty 
signs. We do this by simply enqueuing the tile entity packets to 
the same asynchronous processor threads as the MAP_CHUNK and 
MAP_CHUNK_BULK packets, causing them to be sent in the same order 
as they were enqueued.